Refactor PersistentCredentials
Try to improve code readability by introducing a callback method for
the interaction between the SingleUseCredentialsProvider and the
PersistentCredentials class.

Improve thread safety by making the getUsernameAndPassword method
syncronized. This avoids to show multiple dialogues in parallel to
the user in parallel.
